Correção dados do banco
Correção de dados de teste para dados reais.

+++ lina/src/main/resources/db/migration/V1__criar_schema_inicial.sql
@@ -83,7 +83,7 @@ CREATE TABLE item_cardapio (
 CREATE TABLE nutricionista (
                                id                      BIGSERIAL PRIMARY KEY,
                                atendimentos_realizados INTEGER NOT NULL,
-                               avaliacao               DOUBLE PRECISION NOT NULL,
+                               avaliacao               INTEGER NOT NULL,
                                endereco_id             BIGINT REFERENCES endereco(id),
                                crn                     VARCHAR(255),
                                email                   VARCHAR(255),
